requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
catboost>=1.2.0
autogluon.timeseries>=1.1.0
torch-geometric>=2.5.0
transformers>=4.35.0
fredapi>=0.5.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
pyyaml>=6.0
tqdm>=4.66.0
